+++ packages/bot/bot-factory-client/src/index.ts
@@ -3,4 +3,3 @@
 //
 
 export * from './bot-factory-client';
-export * from './generate-invitation';